+++ download-video-gui.sh
@@ -3,8 +3,8 @@
 # SPDX-License-Identifier: MIT
 # ============================================================================
 # Name        : download-video-gui.sh
-# Version     : 2.1.20
-# Date        : 2026-08-02
+# Version     : 2.1.21
+# Date        : 2026-08-20
 # Description : Zenity GUI for MKV video, authenticated YouTube HLS, or audio.
 # ============================================================================